(maint) Fix parent directory generation for subdirectories
When generating indices for a subset of the release-archives site (apt),
I noticed the 'Parent Directory' for nested directories wasn't working
quite as expected. For example, with
'http://release-archives.puppet.com/apt/dists/' the Parent Directory
shows up as 'http://release-archives.puppet.com/index_by_name.html'.

When kicking off the job both the prefix and the base_path were
'release-archives.puppet.com/apt'. Since this is a nested directory,
when stripping the base_path from the prefix, we only want to remove the
base_path up to the first '/'.