(maint) Add new optional positional parameter for the base_path to strip
In general cases, we're generating indexes for the top level of a site,
so the base_path we need to strip out from URLs is the same as the path
we start traversing. However, if we're generating indexes for a
subdirectory, we need to be able to specify what part of the URLs need
to be stripped.